% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundsurfeR_functions.R
\name{bma}
\alias{bma}
\title{bma}
\usage{
bma(wav_file, tempbin)
}
\arguments{
\item{class}{binTemplateList , object created by combining a list of binary templates}
}
\description{
This function allows the user to go through the entire analysis of binary match point using a single line of code
}
\examples{
# sample code
}
\keyword{binary,}
\keyword{matching,}
\keyword{template}
\keyword{wave,}
