% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundsurfeR_functions.R
\name{cma}
\alias{cma}
\title{cma}
\usage{
cma(wav_file, tempcor)
}
\arguments{
\item{class}{corTemplateList , object created by combining a list of binary templates}
}
\description{
This function allows the user to go through the entire process cross correlation template analysis using a single line of code
}
\examples{
# sample code
}
\keyword{correlation,}
\keyword{cross,}
\keyword{matching,}
\keyword{template}
\keyword{wave,}
