% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundsurfeR_functions.R
\name{demean}
\alias{demean}
\title{demean}
\usage{
demean(x)
}
\arguments{
\item{signal}{extracted from wave object}
}
\description{
This function allows the user to remove DC offset
}
\keyword{DC}
\keyword{offset,}
\keyword{punny}
\keyword{signal,}
\keyword{wave,}
