% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundsurfer-package.R
\docType{package}
\name{soundsurfer-package}
\alias{soundsurfer}
\alias{soundsurfer-package}
\title{soundsurfer: Automatic Acoustic Template Matching}
\description{
The functions included in this package lets you conduct acoustic visulization and analysis including feature extraction and identification with .wav data files.
}
\keyword{internal}
